Add M-TDPC to simulation and change sign of force
Adding M-TDPC meant reevaluating the sign of the force used in the model. For reasons I cannot recall, I made the therapist force the negative of what would eventually be applied to the patient. By changing the sign of the error calculation in the Therapist subsystem and changing the sign again in the subtract block immediately before the patient subsystem, I made the entire simulation use a consistent sign convention for force.